% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTraces.R
\name{addTraces}
\alias{addTraces}
\title{add 3D scatterplot traces to a plotly plot, each corresponding to a dropdown}
\usage{
addTraces(cols, dimcols, shared)
}
\arguments{
\item{cols}{the names of the grouping columns in the data}

\item{dimcols}{the names of the dimensions columns in the data}

\item{shared}{an R data.frame or a crosstalk::SharedData object}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       a plotly plot object with at least one 3D plot trace
}\if{html}{\out{</div>}}
}
\description{
add 3D scatterplot traces to a plotly plot, each corresponding to a dropdown
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      plotdfly
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{      choosePalette
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{      mergePalettes
}\if{html}{\out{</div>}}
}
