% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabData.R
\name{grabData}
\alias{grabData}
\title{Let plotUMAPly handle Seurat, SingleCellExperiment, or data.frame objects}
\usage{
grabData(x, use = "UMAP", color_by = "group", label_by = NULL, dims = 1:2)
}
\arguments{
\item{x}{an object (a data.frame, RcppML::nmf, SCE, or Seurat)}

\item{use}{name of dimred to use, if x is not a df or matrix ("UMAP")}

\item{color_by}{column name of metadata (or x itself) to color by ("study")}

\item{label_by}{column name of metadata (or of x itself) to label by (NULL)}

\item{dims}{what dimred columns (or plain old columns) in x to use (1:6)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a data.frame
}\if{html}{\out{</div>}}
}
\description{
Let plotUMAPly handle Seurat, SingleCellExperiment, or data.frame objects
}
