% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.asr.R
\name{inv.asr}
\alias{inv.asr}
\title{inverse arcsine square root (inv.asr) recovers asr'ed proportions in (0, 1)}
\usage{
inv.asr(p)
}
\arguments{
\item{x}{a nonnegative transformed proportion}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ the original proportion p, IFF x = asr(p) and 0 <= p <= 1
}\if{html}{\out{</div>}}
}
\description{
the inverse (asr) is provided for completeness; reconstruction error is
typically in the range of 1e-3 to 1e-9 where err = abs(x - asr(inv.asr(x)))
assuming that 0 <= x <= 1.57079 or thereabouts (i.e., asr(0) <= x <= asr(1)).
}
\examples{
x0 <- c(0, 0.032, 0.1, 0.32, 0.79, 1.25, 1.47, 1.54, 1.57079)
p <- inv.asr(x)
max(abs(x0 - asr(p))) 

}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{asr
}\if{html}{\out{</div>}}
}
