% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDropdown.R
\name{makeDropdown}
\alias{makeDropdown}
\title{create dropdown lists for changing aspects of a plot_ly plot}
\usage{
makeDropdown(cols, shared)
}
\arguments{
\item{cols}{the names of the grouping columns in the data}

\item{shared}{an R data.frame or a crosstalk::SharedData object}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       a dropdown menu represented as a list of lists
}\if{html}{\out{</div>}}
}
\description{
create dropdown lists for changing aspects of a plot_ly plot
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      plotdfly
}\if{html}{\out{</div>}}
}
