% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePalettes.R
\name{mergePalettes}
\alias{mergePalettes}
\title{specifically for plot_ly multi-grouping plots, merge multiple palettes}
\usage{
mergePalettes(pals, ...)
}
\arguments{
\item{pals}{a list of palettes to merge}

\item{...}{not currently used}
}
\description{
After generating (or retrieving cached) palettes of the appropriate size,
it turns out that one must merge them for plot_ly to use multiple groups.
Instead of generating an enormous and possibly convergent palette, it is
usually better to generate individual grouping-specific palettes, and then
merge them. This function handles the latter and resolves any collisions.
Ties (same name in different groupings) are broken by using whichever color
was seen first for that name. In extreme cases this may result in poor
divergence of colors for duplicated names, so try not to have any.
}
\examples{

data("defaultPalettes")
pals <- defaultPalettes[c(1, 10, 100)]
mergePalettes(pals)

}
\seealso{
choosePalette

getPalette
}
