% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numTraces.R
\name{numTraces}
\alias{numTraces}
\title{compute the number of traces that could be laid out in a plotly plot}
\usage{
numTraces(m, d = 3, g = 1)
}
\arguments{
\item{m}{the total number of columns (dimensions) to support}

\item{d}{number of dimensions to plot simultaneously (3)}

\item{g}{the number of groupings to support (1)}
}
\value{
how many traces this implies
}
\description{
compute the number of traces that could be laid out in a plotly plot
}
\examples{

numTraces(m=3, d=3, g=1) # 6     (i.e. xyz, yzx, zxy, xzy, zyx, yxz)
numTraces(m=6, d=3, g=3) # 360
numTraces(m=4, d=3, g=8) # 192
numTraces(m=8, d=3, g=4) # 1344

}
