% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP2D.R
\name{plotUMAP2D}
\alias{plotUMAP2D}
\title{toss a 2D UMAP plot up on the interwebs with plotly}
\usage{
plotUMAP2D(
  x,
  use = "UMAP",
  color_by = "group",
  label_by = NULL,
  static = TRUE,
  shuffle = FALSE,
  ...
)
}
\arguments{
\item{x}{an object (a data.frame, matrix, SCE, or Seurat)}

\item{use}{name of dimred to use, if x is not a df or matrix ("UMAP")}

\item{color_by}{column name of metadata (or x itself) to color by ("group")}

\item{label_by}{column name of metadata (or x itself) to label by ("group")}

\item{static}{use cached palette(s) of the appropriate size? (TRUE)}

\item{shuffle}{shuffle the colors of the cached palette(s)? (FALSE)}

\item{...}{other stuff passed to plotly}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     nothing much; a static webpage is created as a side effect.
}\if{html}{\out{</div>}}
}
\description{
toss a 2D UMAP plot up on the interwebs with plotly
}
