% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAPly.R
\name{plotUMAPly}
\alias{plotUMAPly}
\title{toss a 3D UMAP plot up on the interwebs with plotly}
\usage{
plotUMAPly(
  x,
  use = "UMAP",
  color_by = "study",
  label_by = NULL,
  dims = 1:3,
  static = TRUE,
  shuffle = FALSE,
  ...
)
}
\arguments{
\item{x}{an object (a data.frame, matrix, SCE, or Seurat)}

\item{use}{name of dimred to use, if x is not a df or matrix ("UMAP")}

\item{color_by}{column name of metadata (or x itself) to color by ("study")}

\item{label_by}{column name of metadata (or of x itself) to label by (NULL)}

\item{dims}{what dimred columns (or plain old columns) in x to use (1:3)}

\item{static}{use cached palette(s) of the appropriate size? (TRUE)}

\item{shuffle}{shuffle the colors of the cached palette(s)? (FALSE)}

\item{...}{other stuff passed to plotly}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     nothing much; a static webpage is created as a side effect.
}\if{html}{\out{</div>}}
}
\description{
TODO
\enumerate{
\item add lines for observation-observation (cell) interactions (e.g. ligands)
\item add switching capability (data points / colors) (Ember has done this now)
}
}
