% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatData.R
\name{whatData}
\alias{whatData}
\title{determine what kind of object grabData should expect to process and how}
\usage{
whatData(x, use = "UMAP", color_by = "study", label_by = NULL, dims = 1:3)
}
\arguments{
\item{x}{an object (a data.frame, matrix, SCE, or Seurat)}

\item{use}{name of dimred to use, if x is not a df or matrix ("UMAP")}

\item{color_by}{column name of metadata (or x itself) to color by ("study")}

\item{label_by}{column name of metadata (or of x itself) to label by (NULL)}

\item{dims}{what dimred columns (or plain old columns) in x to use (1:3)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a string with the (sanity checked) data type, or an error
}\if{html}{\out{</div>}}
}
\description{
Currently we try data.frame, SingleCellExperiment, and Seurat, in that order
}
